requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
python-dotenv==1.0.0
streamlit==1.28.1
requests==2.31.0
Pillow==10.1.0
huggingface-hub==0.19.4
google-generativeai==0.3.2
langchain==0.1.0
langchain-community==0.0.10
langchain-chroma==0.1.4
langchain-huggingface==0.0.1
pypdf==3.17.4
chromadb==0.4.18
sentence-transformers==2.2.2
pydantic==2.5.0